group creation & message posting for principal, S3 bucket updated

+++ app/src/main/java/com/aanglearning/principalapp/dao/GroupDao.java
@@ -18,21 +18,25 @@ import java.util.List;
 public class GroupDao {
 
     public static int insert(Groups group) {
-        String sql = "insert into groups(Id, Name, SectionId, IsSection, ClassId, IsClass, CreatedBy, CreatedDate, IsActive) " +
-                "values(?,?,?,?,?,?,?,?,?)";
+        String sql = "insert into groups(Id, Name, IsSchool, SectionId, IsSection, ClassId, IsClass, " +
+                "CreatedBy, CreatorName, CreatorRole, CreatedDate, IsActive) " +
+                "values(?,?,?,?,?,?,?,?,?,?,?,?)";
         SQLiteDatabase db = AppGlobal.getSqlDbHelper().getWritableDatabase();
         db.beginTransactionNonExclusive();
         SQLiteStatement stmt = db.compileStatement(sql);
         try {
             stmt.bindLong(1, group.getId());
             stmt.bindString(2, group.getName());
-            stmt.bindLong(3, group.getSectionId());
-            stmt.bindString(4, Boolean.toString(group.isSection()));
-            stmt.bindLong(5, group.getClassId());
-            stmt.bindString(6, Boolean.toString(group.isClas()));
-            stmt.bindLong(7, group.getCreatedBy());
-            stmt.bindString(8, group.getCreatedDate());
-            stmt.bindString(9, Boolean.toString(group.isActive()));
+            stmt.bindString(3, Boolean.toString(group.isSchool()));
+            stmt.bindLong(4, group.getSectionId());
+            stmt.bindString(5, Boolean.toString(group.isSection()));
+            stmt.bindLong(6, group.getClassId());
+            stmt.bindString(7, Boolean.toString(group.isClas()));
+            stmt.bindLong(8, group.getCreatedBy());
+            stmt.bindString(9, group.getCreatorName());
+            stmt.bindString(10, group.getCreatorRole());
+            stmt.bindString(11, group.getCreatedDate());
+            stmt.bindString(12, Boolean.toString(group.isActive()));
             stmt.executeInsert();
             stmt.clearBindings();
         } catch (Exception e) {
@@ -45,8 +49,9 @@ public class GroupDao {
     }
 
     public static int insertMany(List<Groups> groups) {
-        String sql = "insert into groups(Id, Name, SectionId, IsSection, ClassId, IsClass, CreatedBy, CreatedDate, IsActive) " +
-                "values(?,?,?,?,?,?,?,?,?)";
+        String sql = "insert into groups(Id, Name, IsSchool, SectionId, IsSection, ClassId, IsClass, " +
+                "CreatedBy, CreatorName, CreatorRole, CreatedDate, IsActive) " +
+                "values(?,?,?,?,?,?,?,?,?,?,?,?)";
         SQLiteDatabase db = AppGlobal.getSqlDbHelper().getWritableDatabase();
         db.beginTransactionNonExclusive();
         SQLiteStatement stmt = db.compileStatement(sql);
@@ -54,13 +59,16 @@ public class GroupDao {
             for(Groups group: groups) {
                 stmt.bindLong(1, group.getId());
                 stmt.bindString(2, group.getName());
-                stmt.bindLong(3, group.getSectionId());
-                stmt.bindString(4, Boolean.toString(group.isSection()));
-                stmt.bindLong(5, group.getClassId());
-                stmt.bindString(6, Boolean.toString(group.isClas()));
-                stmt.bindLong(7, group.getCreatedBy());
-                stmt.bindString(8, group.getCreatedDate());
-                stmt.bindString(9, Boolean.toString(group.isActive()));
+                stmt.bindString(3, Boolean.toString(group.isSchool()));
+                stmt.bindLong(4, group.getSectionId());
+                stmt.bindString(5, Boolean.toString(group.isSection()));
+                stmt.bindLong(6, group.getClassId());
+                stmt.bindString(7, Boolean.toString(group.isClas()));
+                stmt.bindLong(8, group.getCreatedBy());
+                stmt.bindString(9, group.getCreatorName());
+                stmt.bindString(10, group.getCreatorRole());
+                stmt.bindString(11, group.getCreatedDate());
+                stmt.bindString(12, Boolean.toString(group.isActive()));
                 stmt.executeInsert();
                 stmt.clearBindings();
             }
@@ -73,27 +81,6 @@ public class GroupDao {
         return 1;
     }
 
-    public static Groups getGroup(long id) {
-        Groups group = new Groups();
-        SQLiteDatabase sqliteDatabase = AppGlobal.getSqlDbHelper().getReadableDatabase();
-        Cursor c = sqliteDatabase.rawQuery("select * from groups where id = " + id, null);
-        c.moveToFirst();
-        while (!c.isAfterLast()) {
-            group.setId(c.getLong(c.getColumnIndex("Id")));
-            group.setName(c.getString(c.getColumnIndex("Name")));
-            group.setSectionId(c.getLong(c.getColumnIndex("SectionId")));
-            group.setSection(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsSection"))));
-            group.setClassId(c.getLong(c.getColumnIndex("ClassId")));
-            group.setClas(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsClass"))));
-            group.setCreatedBy(c.getLong(c.getColumnIndex("CreatedBy")));
-            group.setCreatedDate(c.getString(c.getColumnIndex("CreatedDate")));
-            group.setActive(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsActive"))));
-            c.moveToNext();
-        }
-        c.close();
-        return group;
-    }
-
     public static List<Groups> getGroups() {
         List<Groups> groups = new ArrayList<>();
         SQLiteDatabase sqliteDatabase = AppGlobal.getSqlDbHelper().getReadableDatabase();
@@ -103,11 +90,40 @@ public class GroupDao {
             Groups group = new Groups();
             group.setId(c.getLong(c.getColumnIndex("Id")));
             group.setName(c.getString(c.getColumnIndex("Name")));
+            group.setSchool(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsSchool"))));
             group.setSectionId(c.getLong(c.getColumnIndex("SectionId")));
             group.setSection(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsSection"))));
             group.setClassId(c.getLong(c.getColumnIndex("ClassId")));
             group.setClas(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsClass"))));
             group.setCreatedBy(c.getLong(c.getColumnIndex("CreatedBy")));
+            group.setCreatorName(c.getString(c.getColumnIndex("CreatorName")));
+            group.setCreatorRole(c.getString(c.getColumnIndex("CreatorRole")));
+            group.setCreatedDate(c.getString(c.getColumnIndex("CreatedDate")));
+            group.setActive(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsActive"))));
+            groups.add(group);
+            c.moveToNext();
+        }
+        c.close();
+        return groups;
+    }
+
+    public static List<Groups> getPrincipalGroups(long teacherId) {
+        List<Groups> groups = new ArrayList<>();
+        SQLiteDatabase sqliteDatabase = AppGlobal.getSqlDbHelper().getReadableDatabase();
+        Cursor c = sqliteDatabase.rawQuery("select * from groups where CreatorRole = 'principal' and CreatedBy = " + teacherId, null);
+        c.moveToFirst();
+        while (!c.isAfterLast()) {
+            Groups group = new Groups();
+            group.setId(c.getLong(c.getColumnIndex("Id")));
+            group.setName(c.getString(c.getColumnIndex("Name")));
+            group.setSchool(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsSchool"))));
+            group.setSectionId(c.getLong(c.getColumnIndex("SectionId")));
+            group.setSection(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsSection"))));
+            group.setClassId(c.getLong(c.getColumnIndex("ClassId")));
+            group.setClas(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsClass"))));
+            group.setCreatedBy(c.getLong(c.getColumnIndex("CreatedBy")));
+            group.setCreatorName(c.getString(c.getColumnIndex("CreatorName")));
+            group.setCreatorRole(c.getString(c.getColumnIndex("CreatorRole")));
             group.setCreatedDate(c.getString(c.getColumnIndex("CreatedDate")));
             group.setActive(Boolean.parseBoolean(c.getString(c.getColumnIndex("IsActive"))));
             groups.add(group);
@@ -120,7 +136,17 @@ public class GroupDao {
     public static int clear() {
         SQLiteDatabase sqliteDb = AppGlobal.getSqlDbHelper().getWritableDatabase();
         try {
-            sqliteDb.execSQL("delete from groups");
+            sqliteDb.execSQL("delete from groups where CreatorRole != 'principal'");
+        } catch(SQLException e) {
+            return 0;
+        }
+        return 1;
+    }
+
+    public static int clearPrincipalGroup() {
+        SQLiteDatabase sqliteDb = AppGlobal.getSqlDbHelper().getWritableDatabase();
+        try {
+            sqliteDb.execSQL("delete from groups where CreatorRole = 'principal'");
         } catch(SQLException e) {
             return 0;
         }
